+++ mmdet/api/__init__.py
@@ -1,0 +1,4 @@
+from .train import train_detector
+from .inference import inference_detector
+
+__all__ = ['train_detector', 'inference_detector']
