+++ mmdet/api/inference.py
@@ -1,0 +1,54 @@
+import mmcv
+import numpy as np
+import torch
+
+from mmdet.datasets import to_tensor
+from mmdet.datasets.transforms import ImageTransform
+from mmdet.core import get_classes
+
+
+def _prepare_data(img, img_transform, cfg, device):
+    ori_shape = img.shape
+    img, img_shape, pad_shape, scale_factor = img_transform(
+        img, scale=cfg.data.test.img_scale)
+    img = to_tensor(img).to(device).unsqueeze(0)
+    img_meta = [
+        dict(
+            ori_shape=ori_shape,
+            img_shape=img_shape,
+            pad_shape=pad_shape,
+            scale_factor=scale_factor,
+            flip=False)
+    ]
+    return dict(img=[img], img_meta=[img_meta])
+
+
+def inference_detector(model, imgs, cfg, device='cuda:0'):
+
+    imgs = imgs if isinstance(imgs, list) else [imgs]
+    img_transform = ImageTransform(
+        **cfg.img_norm_cfg, size_divisor=cfg.data.test.size_divisor)
+    model = model.to(device)
+    model.eval()
+    for img in imgs:
+        img = mmcv.imread(img)
+        data = _prepare_data(img, img_transform, cfg, device)
+        with torch.no_grad():
+            result = model(**data, return_loss=False, rescale=True)
+        yield result
+
+
+def show_result(img, result, dataset='coco', score_thr=0.3):
+    class_names = get_classes(dataset)
+    labels = [
+        np.full(bbox.shape[0], i, dtype=np.int32)
+        for i, bbox in enumerate(result)
+    ]
+    labels = np.concatenate(labels)
+    bboxes = np.vstack(result)
+    mmcv.imshow_det_bboxes(
+        img.copy(),
+        bboxes,
+        labels,
+        class_names=class_names,
+        score_thr=score_thr)
